# swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "defaultTitle",
    "description" : "defaultDescription",
    "version" : "0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:3000"
  } ],
  "paths" : {
...
    "/products" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "post" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_2"
              },
              "examples" : {
                "0" : {
                  "value" : "{\n    \"name\": \"FifthProductTest\",\n    \"display_name\": \"FifthProductTestDisplayName\",\n    \"description\": \"FifthProductTestDescription\",\n    \"category\": \"FifthProductTestCategory\",\n    \"id\": 5\n}"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    },
    "/categories" : {
      "get" : {
        "description" : "Auto generated using Swagger Inspector",
        "responses" : {
          "200" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "post" : {
        "description" : "Auto generated using Swagger Inspector",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/body_3"
              },
              "examples" : {
                "0" : {
                  "value" : "{\n      \"name\": \"4thTest\",\n      \"display_name\": \"4thTestDisplayName\",\n      \"description\": \"4thTestDescription\",\n      \"id\": 4\n    }"
                }
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Auto generated using Swagger Inspector",
            "content" : {
              "application/json; charset=utf-8" : {
                "schema" : {
                  "type" : "string"
                },
                "examples" : { }
              }
            }
          }
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
      "servers" : [ {
        "url" : "http://localhost:3000"
      } ]
    }
  },
  "components" : {
    "schemas" : {
...
      "body_2" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer"
          },
          "display_name" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string"
          }
        }
      },
      "body_3" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "integer"
          },
          "display_name" : {
            "type" : "string"
          }
        }
      }
    }
  }
}